Scan Control Support Code
Added the Code needed to start a scan and monitor the scan. It is set up
as MVC with Vm and camera being the M, The MainWindow being V. Scan
controller controlls the scan process and MCU controls everything. It
still needs a data interface to OpenCV and a Model for the Data. This is
what will be made next